Added frontendUI

+++ backend/controllers/goalController.js
@@ -1,11 +1,14 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 //@desc Get goals
 //@route GET api /api/goals
 //@access Private
 
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find({});
+  const goals = await Goal.find({
+    user: req.user.id,
+  });
   res.status(200).json(goals);
 });
 
@@ -15,6 +18,7 @@ const postGoals = asyncHandler(async (req, res) => {
     throw new Error("Please add a text field");
   }
   const goal = await Goal.create({
+    user: req.user.id,
     text: req.body.text,
   });
   res.status(200).json(goal);
@@ -25,6 +29,18 @@ const updateGoals = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal not found");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+  // Make sure the logged in user marches the foal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -39,6 +55,17 @@ const delGoals = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Unspecified goal to delete");
+  }
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+  // Make sure the logged in user marches the foal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
   }
   await goal.deleteOne();
   res.status(200).json({
